Move hub_online_once to testing utils
