+++ tests/testing_utils.py
@@ -11,9 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import os
 import unittest
 from contextlib import contextmanager
 from functools import lru_cache, wraps
+from unittest import mock
 
 import numpy as np
 import pytest
@@ -39,6 +41,10 @@ from peft.import_utils import (
     is_optimum_available,
     is_torchao_available,
 )
+
+
+# Globally shared model cache used by `hub_online_once`.
+_HUB_MODEL_CACHE = {}
 
 
 torch_device, device_count, memory_allocated_func = get_backend()
@@ -241,3 +247,56 @@ def set_init_weights_false(config_cls, kwargs):
     else:
         kwargs["init_weights"] = False
     return kwargs
+
+
+@contextmanager
+def hub_online_once(model_id: str):
+    """Set env[HF_HUB_OFFLINE]=1 (and patch transformers/hugging_face_hub to think that it was always that way)
+    for model ids that were seen already so that the hub is not contacted twice for the same model id in said context.
+    The cache (`_HUB_MODEL_CACHE`) also tracks the number of cache hits per model id.
+
+    The reason for doing a context manager and not patching specific methods (e.g., `from_pretrained`) is that there
+    are a lot of places (`PeftConfig.from_pretrained`, `get_peft_state_dict`, `load_adapter`, ...) that possibly
+    communicate with the hub to download files / check versions / etc.
+
+    Note that using this context manager can cause problems when used in code sections that access different resources.
+    Example:
+
+    ```
+    def test_something(model_id, config_kwargs):
+        with hub_online_once(model_id):
+            model = ...from_pretrained(model_id)
+            self.do_something_specific_with_model(model)
+    ```
+    It is assumed that `do_something_specific_with_model` is an absract method that is implement by several tests.
+    Imagine the first test simply does `model.generate([1,2,3])`. The second call from another test suite however uses
+    a tokenizer (`AutoTokenizer.from_pretrained(model_id)`) - this will fail since the first pass was online but didn't
+    use the tokenizer and we're now in offline mode and cannot fetch the tokenizer. The recommended workaround is to
+    extend the cache key (`model_id` passed to `hub_online_once` in this case) by something in case the tokenizer is
+    used, so that these tests don't share a cache pool with the tests that don't use a tokenizer.
+    """
+    global _HUB_MODEL_CACHE
+    override = {}
+
+    try:
+        if model_id in _HUB_MODEL_CACHE:
+            override = {"HF_HUB_OFFLINE": "1"}
+            _HUB_MODEL_CACHE[model_id] += 1
+        else:
+            if model_id not in _HUB_MODEL_CACHE:
+                _HUB_MODEL_CACHE[model_id] = 0
+        with (
+            # strictly speaking it is not necessary to set the environment variable since most code that's out there
+            # is evaluating it at import time and we'd have to reload the modules for it to take effect. It's
+            # probably still a good idea to have it if there's some dynamic code that checks it.
+            mock.patch.dict(os.environ, override),
+            mock.patch("huggingface_hub.constants.HF_HUB_OFFLINE", override.get("HF_HUB_OFFLINE", False) == "1"),
+            mock.patch("transformers.utils.hub._is_offline_mode", override.get("HF_HUB_OFFLINE", False) == "1"),
+        ):
+            yield
+    except Exception:
+        # in case of an error we have to assume that we didn't access the model properly from the hub
+        # for the first time, so the next call cannot be considered cached.
+        if _HUB_MODEL_CACHE.get(model_id) == 0:
+            del _HUB_MODEL_CACHE[model_id]
+        raise
